:recycle: refactor: 예외처리 메서드 Validator로 분리
